Develop

+++ controllers/card.js
@@ -1,16 +1,15 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const ValidationError = require('../errors/validation-error');
+const NotEnoughRights = require('../errors/not-enough-rights');
+const { validationError, cardNotFound, noRightsToRem, castErr } = require('../errors/error-messages');
 
 // Получить список всех карточек
 
 const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch((err) => {
-      next({
-        message: err.message,
-        status: 500
-      });
-    });
+    .catch(next);
 };
 
 // Создание карточки
@@ -22,10 +21,11 @@ const createCard = (req, res, next) => {
   Card.create({ name, link, owner, createdAt })
     .then((card) => res.send(card))
     .catch((err) => {
-      next({
-        message: err.message,
-        status: err.name === 'ValidationError' ? 400 : 500
-      });
+      if (err.name === 'ValidationError') {
+        next(new ValidationError(validationError));
+        return;
+      }
+      next(err);
     });
 };
 
@@ -36,27 +36,22 @@ const deleteCard = (req, res, next) => {
   Card.findById(cardId)
     // eslint-disable-next-line consistent-return
     .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
       if (card.owner.toString() !== req.user._id) {
-        next({
-          message: 'Недостаточно прав, чтобы удалить карточку',
-          status: 403
-        });
-        return;
+        throw new NotEnoughRights(noRightsToRem);
       }
       Card.findByIdAndRemove(cardId)
         .then(() => res.send(card))
-        .catch((err) => {
-          next({
-            message: err.message,
-            status: 500
-          });
-        });
+        .catch(next);
     })
     .catch((err) => {
-      next({
-        message: err.name === 'CastError' ? 'Карточка не существует' : err.message,
-        status: err.name === 'CastError' ? 404 : 500
-      });
+      if (err.name === 'CastError') {
+        next(new ValidationError(castErr));
+        return;
+      }
+      next(err);
     });
 };
 
@@ -64,12 +59,18 @@ const deleteCard = (req, res, next) => {
 
 const setLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send(card))
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
+      res.send(card);
+    })
     .catch((err) => {
-      next({
-        message: err.name === 'CastError' ? 'Карточка не существует' : err.message,
-        status: err.name === 'CastError' ? 404 : 500
-      });
+      if (err.name === 'CastError') {
+        next(new ValidationError(castErr));
+        return;
+      }
+      next(err);
     });
 };
 
@@ -77,12 +78,18 @@ const setLike = (req, res, next) => {
 
 const removeLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send(card))
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
+      res.send(card);
+    })
     .catch((err) => {
-      next({
-        message: err.name === 'CastError' ? 'Карточка не существует' : err.message,
-        status: err.name === 'CastError' ? 404 : 500
-      });
+      if (err.name === 'CastError') {
+        next(new ValidationError(castErr));
+        return;
+      }
+      next(err);
     });
 };
